+++ oportuni/src/app/pages/search/page.tsx
@@ -1,11 +1,11 @@
 "use client"
-import Footer from '@/components/Footer';
+import Footer from '@/src/components/Footer';
 
-const Home = () => {
+const Search = () => {
   return (
     <div className="flex flex-col justify-between min-h-screen w-full bg-white">
       <main className="flex-grow flex items-center justify-center">
-        <h1 className="text-4xl font-bold">Bem-vindo à Página Inicial</h1>
+        <h1 className="text-4xl font-bold">Bem-vindo à Pesquisa</h1>
       </main>
 
       <Footer /> {/* Reusable Footer component */}
@@ -13,4 +13,4 @@ const Home = () => {
   );
 };
 
-export default Home;
+export default Search;
