+++ oportuni/src/pages/api/auth/password/login.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import dbConnect from "@/lib/mongodb";
-import User from "@/models/User";
+import dbConnect from "@/src/lib/mongodb";
+import User from "@/src/models/User";
 import bcrypt from "bcryptjs";
 export default async function handler(
   req: NextApiRequest,
